Add third-party asset licenses
As part of this, we discovered that the carrot chop sound is CC-BY licensed. We don't currently have a way to give attribution, not even in our designs; that is, we have no concrete plans for title screens, credits screens, etc. The safer thing for us is to replace it by something we own or that is public domain. Gavin is working on that now.